WIP: preliminary code transfer from the server to the job; needs to parse invalid errors from the object into an array
